# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin API",
    "description": "Digital Twin - Car Health Monitoring Platform backend. Phase 1 provides vehicle registration and raw telemetry storage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/rag/documents": {
      "post": {
        "tags": [
          "Admin RAG"
        ],
        "summary": "Upload a manufacturer document (admin only)",
        "description": "Validates the multipart file and metadata, then runs it through the existing RAG ingestion pipeline (Docling/text parsing, structure-aware chunking, BGE-M3 embeddings, pgvector). Supported document formats follow the parser registry: .pdf .docx .doc .rtf .pptx .ppt .html .mhtml .xlsx .xls .txt .md .text.",
        "operationId": "upload_document_api_v1_admin_rag_documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_admin_rag_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRagUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin RAG"
        ],
        "summary": "List corpus documents (admin only)",
        "description": "Paginated corpus documents (newest first) with version counts. Optional exact filters: make, model, canonical; model-year range via year; documents having a version with the given ingestion status.",
        "operationId": "list_documents_api_v1_admin_rag_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Make"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "canonical",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Canonical"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 16
                },
                {
                  "type": "null"
                }
              ],
              "description": "Version ingestion status (e.g. completed)",
              "title": "Status"
            },
            "description": "Version ingestion status (e.g. completed)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RAGDocumentSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rag/documents/{document_id}": {
      "get": {
        "tags": [
          "Admin RAG"
        ],
        "summary": "Get document detail (admin only)",
        "description": "Document identity metadata plus its immutable version history (content hash, parser, chunk counts, ingestion status per version).",
        "operationId": "get_document_api_v1_admin_rag_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRagDocumentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin RAG"
        ],
        "summary": "Delete a document (admin only)",
        "description": "Transactionally removes the document, all of its immutable versions and every chunk/vector row. RAG search stops returning its content immediately.",
        "operationId": "delete_document_api_v1_admin_rag_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminRagDocumentDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "canonical_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Source"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_year_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Year Start"
          },
          "model_year_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Year End"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count",
            "default": 0
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/RAGVersionSummary"
            },
            "type": "array",
            "title": "Versions"
          },
          "latest_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RAGVersionSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type",
          "document_type",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "AdminRagDocumentDetailResponse",
        "description": "One document plus its immutable version history (each with chunk count)."
      },
      "AdminRagUploadResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ingested",
              "unchanged"
            ],
            "title": "Status"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "canonical": {
            "type": "string",
            "title": "Canonical"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "chunks_created": {
            "type": "integer",
            "title": "Chunks Created",
            "default": 0
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash",
            "default": ""
          },
          "parser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Name"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "embedding_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dimension"
          }
        },
        "type": "object",
        "required": [
          "status",
          "filename",
          "canonical"
        ],
        "title": "AdminRagUploadResponse",
        "description": "Outcome of one admin upload, using the existing ingestion semantics.\n\n``status`` is ``ingested`` (a new document/version was persisted) or\n``unchanged`` (identical content already completed \u2014 Phase 5 idempotency).\nA failed ingestion raises instead of returning a fabricated success."
      },
...
      "Body_upload_document_api_v1_admin_rag_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Document file to ingest"
          },
          "canonical": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Canonical"
          },
          "make": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "year_end": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year End"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "file",
          "canonical"
        ],
        "title": "Body_upload_document_api_v1_admin_rag_documents_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_RAGDocumentSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RAGDocumentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginatedResponse[RAGDocumentSummary]"
      },
...
      "RAGDocumentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "canonical_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Source"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_year_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Year Start"
          },
          "model_year_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Year End"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type",
          "document_type",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "RAGDocumentSummary",
        "description": "Read-only admin view of one document in the corpus."
      },
...
      "RAGVersionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "parser_name": {
            "type": "string",
            "title": "Parser Name"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "embedding_dimension": {
            "type": "integer",
            "title": "Embedding Dimension"
          },
          "chunking_version": {
            "type": "string",
            "title": "Chunking Version"
          },
          "ingestion_status": {
            "type": "string",
            "title": "Ingestion Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "version",
          "content_hash",
          "source_filename",
          "parser_name",
          "embedding_model",
          "embedding_dimension",
          "chunking_version",
          "ingestion_status",
          "created_at"
        ],
        "title": "RAGVersionSummary",
        "description": "Read-only admin view of one immutable document version."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}